requirements.txt
# Core PDF processing
PyMuPDF>=1.23.0

# Image processing for logo detection
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# CLI framework
click>=8.1.0

# Configuration
pyyaml>=6.0
pydantic>=2.0.0

# Progress and UI
tqdm>=4.65.0
colorama>=0.4.6
